Rank 밑에 key 표시하게 변경, gameGroup 관련 명령어 작업중 TODO: 명령어 JSON 손볼것

+++ main/java/kuro9/mahjongbot/db/data/Game.kt
@@ -9,6 +9,6 @@ data class Game(
     @UserRes val addedBy: Long,
     val gameGroup: String = ""
 ) {
-    val id: Long = 0
+    val id: Int = 0
     val createdAt: Timestamp = Timestamp(0)
 }
